+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Created `Stoplight::Notifier` and subclasses.
 - Sent notifications when moving from green to red.
+- Renamed `Stoplight::Light::DEFAULT_THRESHOLD` to
+  `Stoplight::DEFAULT_THRESHOLD`.
 - Renamed `Stoplight::Error::NoFallback` to `Stoplight::Error::RedLight`.
 - Created `Stoplight::Mixin#stoplight` for easily creating and running simple
   stoplights.